Fix: tanstack query 설정 로직 분리
