+++ apps/web/app/providers.tsx
@@ -1,36 +1,13 @@
 "use client";
 
-import { isServer, QueryClient, QueryClientProvider } from "@tanstack/react-query";
+import { QueryClientProvider } from "@tanstack/react-query";
 import { ReactQueryDevtools } from "@tanstack/react-query-devtools";
-import { type ReactNode } from "react";
+import { type PropsWithChildren } from "react";
 import { Toast } from "@ui/index";
 import MobileSizeWatcher from "@/components/MobileSizeWatcher";
+import { getQueryClient } from "@/lib/tanstackQuery";
 
-interface ProvidersProps {
-  children: ReactNode;
-}
-
-function makeQueryClient(): QueryClient {
-  return new QueryClient({
-    defaultOptions: {
-      queries: {
-        staleTime: 60 * 1000,
-      },
-    },
-  });
-}
-
-let browserQueryClient: QueryClient | undefined;
-
-function getQueryClient(): QueryClient {
-  if (isServer) {
-    return makeQueryClient();
-  }
-  if (!browserQueryClient) browserQueryClient = makeQueryClient();
-  return browserQueryClient;
-}
-
-export function Providers({ children }: ProvidersProps): JSX.Element {
+export function Providers({ children }: PropsWithChildren): JSX.Element {
   const queryClient = getQueryClient();
 
   return (
